Fix Monoid instance for Color type
I chose the original position of the args, as it felt natural to me that
mappending more on the end was like stacking more on the top. However,
given the associativity of mappend, this precludes any short-circuiting
when a color is encountered. The new method means that when a color is
encountered, everything mappended after it will remain unevaluated, as
we can't see through colors (in this simple, terminal color scheme).