Add old checkswirl.ppm, and recreation
Added under original PPM's last modified date via:
stat -c %y -> git commit --date="..."

Date was:
2017-11-26 23:10:44.859164535 -0800

Recreation necessary as the original code was created interactively in
GHCi, and never saved. It took quite a bit of playing to find the
numbers again, and I'm not sure it's 100% accurate, but it's very close.

The recreation code is now in examples/PPM/ExamplesPPM.hs.